+++ nixio/pycore/__init__.py
@@ -1,10 +1,8 @@
 from .property import Property
 from .section import Section
-from .source import Source
 from .data_array import DataArray
 from .dimensions import RangeDimension, SetDimension, SampledDimension
 from .feature import Feature
 from .tag import Tag
 from .multi_tag import MultiTag
 from .group import Group
-from .block import Block
